policy admin and cashier

+++ app/Http/Controllers/API/CategoryController.php
@@ -5,38 +5,38 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\CategoryResource;
 use App\Models\Category;
+use App\Models\User;
 use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 
 class CategoryController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
-       $categories = Category::get();
+        $categories = Category::get();
 
-        if(count($categories) > 0){
-
-           return CategoryResource::collection($categories);
-
-        }else{
-            return response()->json(['message' =>'No data found'], 404);
+        if (count($categories) > 0) {
+            return CategoryResource::collection($categories);
+        } else {
+            return response()->json(['message' => 'No data found'], 404);
         }
     }
 
-
     public function store(Request $request)
     {
+        if (Gate::denies('create', auth()->user())) {
+            return response()->json(['message' => 'Unauthorized.'], 403);
+        }
+
         try {
             $validator = Validator::make($request->all(), [
-                'name' => 'required|string|:255',
+                'name' => 'required|string|max:255',
                 'description' => 'required|string',
-                'image' => 'nullable|image|mimes:jpg,jpeg,png|max:2048', 'image'
+                'image' => 'nullable|image|mimes:jpg,jpeg,png|max:2048'
             ]);
 
             if ($validator->fails()) {
@@ -45,10 +45,10 @@ class CategoryController extends Controller
                     'errors' => $validator->messages(),
                 ], 422);
             }
-            if (request()->hasFile('image')) {
+
+            if ($request->hasFile('image')) {
                 $name = $request->file('image')->getClientOriginalName();
                 $path = $request->file('image')->storeAs('category-photos', $name);
-
             }
 
             $category = Category::create([
@@ -56,53 +56,54 @@ class CategoryController extends Controller
                 'description' => $request->description,
                 'image' => $path ?? '',
             ]);
+
             return response()->json([
                 'message' => 'Category created successfully',
-                'data' => $category
+                'data' => $category,
             ], 200);
-        }catch (Exception $e) {
+        } catch (Exception $e) {
             Log::error('Category creation failed: ' . $e->getMessage());
 
             return response()->json([
                 'message' => 'Category creation failed',
-                'error' => $e->getMessage()
+                'error' => $e->getMessage(),
             ], 500);
         }
     }
 
-
-
-    /**
-     * Display the specified resource.
-     */
     public function show(Category $category)
     {
-        try{
+        try {
             return new CategoryResource($category);
-        }catch(Exception $e){
+        } catch (Exception $e) {
             return response()->json([
                 'message' => 'Category show failed',
-                'error' => $e->getMessage()
-            ],500);
+                'error' => $e->getMessage(),
+            ], 500);
         }
     }
 
-
     public function update(Request $request, Category $category)
     {
+        if (Gate::denies('update', auth()->user())) {
+            return response()->json(['message' => 'Unauthorized.'], 403);
+        }
+
         try {
             $validator = Validator::make($request->all(), [
-                'name' => 'required|string|:255',
+                'name' => 'required|string|max:255',
                 'description' => 'required|string',
-                'image' => 'nullable|image|mimes:jpg,jpeg,png|max:2048', 'image'
+                'image' => 'nullable|image|mimes:jpg,jpeg,png|max:2048'
             ]);
+
             if ($validator->fails()) {
                 return response()->json([
                     'message' => 'Category validation error',
                     'errors' => $validator->messages(),
                 ], 422);
             }
-            if (request()->hasFile('image')) {
+
+            if ($request->hasFile('image')) {
                 if (isset($category->image)) {
                     Storage::delete($category->image);
                 }
@@ -115,26 +116,28 @@ class CategoryController extends Controller
                 'description' => $request->description,
                 'image' => $path ?? $category->image,
             ]);
+
             return response()->json([
-                'message' => 'Category update successfully',
-                'data' => $category
+                'message' => 'Category updated successfully',
+                'data' => $category,
             ], 200);
-        }catch (Exception $e) {
+        } catch (Exception $e) {
             Log::error('Category update failed: ' . $e->getMessage());
 
             return response()->json([
-                'message' => 'Product update failed',
-                'error' => $e->getMessage()
+                'message' => 'Category update failed',
+                'error' => $e->getMessage(),
             ], 500);
         }
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(Category $category)
     {
-        if(isset($category->image)) {
+        if (Gate::denies('delete', auth()->user())) {
+            return response()->json(['message' => 'Unauthorized.'], 403);
+        }
+
+        if (isset($category->image)) {
             Storage::delete($category->image);
         }
         $category->delete();
